requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlmodel>=0.0.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.0.0
Pillow>=10.0.0
pydantic-settings>=2.0.0

